[209] Simplify the project views by sharing common code in the router
Bug: https://github.com/svalyn/svalyn-studio/issues/209
