+++ frontend/svalyn-studio-app/src/projects/changeproposals/ProjectChangeProposalsView.types.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022 Stéphane Bégaudeau.
+ * Copyright (c) 2022, 2023 Stéphane Bégaudeau.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
  * associated documentation files (the "Software"), to deal in the Software without restriction,
@@ -17,14 +17,7 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-export interface ProjectChangeProposalProps {
-  projectIdentifier: string;
-  role: MembershipRole;
-}
-
-export type MembershipRole = 'ADMIN' | 'MEMBER' | 'NONE';
-
-export interface ProjectChangeProposalState {
+export interface ProjectChangeProposalsViewState {
   project: Project | null;
   filter: ChangeProposalStatusFilter;
   selectedChangeProposalIds: string[];
