+++ frontend/svalyn-studio-app/src/projects/useProject.types.ts
@@ -17,23 +17,21 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-import Box from '@mui/material/Box';
-import { useParams } from 'react-router-dom';
-import { Workbench } from '../../workbench/Workbench';
+export interface ProjectContextValue {
+  project: Project;
+}
 
-export const WorkspaceView = () => {
-  const { changeId } = useParams();
-  return (
-    <Box
-      sx={{
-        display: 'grid',
-        gridTemplateRows: 'minmax(0, 1fr)',
-        gridTemplateColumns: 'minmax(0, 1fr)',
-        width: '100vw',
-        height: '100vh',
-      }}
-    >
-      {changeId ? <Workbench changeId={changeId} /> : null}
-    </Box>
-  );
-};
+export interface Project {
+  identifier: string;
+  name: string;
+  description: string;
+  organization: Organization;
+}
+
+export interface Organization {
+  identifier: string;
+  name: string;
+  role: MembershipRole;
+}
+
+export type MembershipRole = 'ADMIN' | 'MEMBER' | 'NONE';
