+++ frontend/svalyn-studio-app/src/workspace/WorkspaceView.types.ts
